% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_chemical_class.R
\name{eval_chemical_class}
\alias{eval_chemical_class}
\title{Determine the parameter's class}
\usage{
eval_chemical_class(value, code_sandre_parameter, Zmoy)
}
\arguments{
\item{value}{parameter value}

\item{code_sandre_parameter}{Sandre's code of the physico-chemical parameter}

\item{Zmoy}{average depth}
}
\value{
the parameter's class
}
\description{
Determine the parameter's class
}
