% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_part_a.R
\name{threshold_part_a}
\alias{threshold_part_a}
\title{Determine the first threshold's vector}
\usage{
threshold_part_a(code_sandre_parameter, Zmoy)
}
\arguments{
\item{code_sandre_parameter}{Sandre's code of the physico-chemical parameter}

\item{Zmoy}{average depth}
}
\value{
the first threshold's vector
}
\description{
Determine the first threshold's vector
}
