% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_part_b.R
\name{threshold_part_b}
\alias{threshold_part_b}
\title{Determine the second threshold's vector}
\usage{
threshold_part_b(code_sandre_parameter, Zmoy)
}
\arguments{
\item{code_sandre_parameter}{Sandre's code of the physico-chemical parameter}

\item{Zmoy}{average depth}
}
\value{
the second threshold's vector
}
\description{
Determine the second threshold's vector
}
