% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_values.R
\name{threshold_values}
\alias{threshold_values}
\title{Select the threshold's vector}
\usage{
threshold_values(code_sandre_parameter, Zmoy)
}
\arguments{
\item{code_sandre_parameter}{Sandre's code of the physico-chemical parameter}

\item{Zmoy}{average depth}
}
\value{
the threshold's vector
}
\description{
Select the threshold's vector
}
\keyword{threshold's}
\keyword{values,}
\keyword{vector}
